% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_mean.R
\name{get_marg_mean}
\alias{get_marg_mean}
\title{Apply a function to MCMC draws of model parameters to give a sample from 'marginal mean'}
\usage{
get_marg_mean(xi_col, eta_col, sigma_error, ui, vi, samples = FALSE)
}
\arguments{
\item{xi_col}{A vector of MCMC draws from the parameter xi}

\item{eta_col}{A vector of MCMC draws from the parameter eta}

\item{sigma_error}{MCMC draws from the parameter `sigma_error`}

\item{ui}{A matrix of draws from random effect (zero part)}

\item{vi}{A matrix of draws from random effect (non-zero part)}

\item{samples}{Return the samples rather than summarizing (default is FALSE)}
}
\value{
Quantiles 0.025, 0.5, 0.975 from a posterior sample of marginal mean
}
\description{
Apply a function to MCMC draws of model parameters to give a sample from 'marginal mean'
}
