% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_mean.R
\name{get_ui_vi}
\alias{get_ui_vi}
\title{Simulate correlated random effects for a new patient}
\usage{
get_ui_vi(nsims, sigma_u, sigma_v, rho)
}
\arguments{
\item{nsims}{The number of simulations used in the inner Monte Carlo integration.}

\item{sigma_u}{MCMC draws from the parameter `sigma_u`}

\item{sigma_v}{MCMC draws from the parameter `sigma_v`}

\item{rho}{MCMC draws from the parameter `rho`}
}
\value{
A list of two matrices, `ui` and `vi`.
}
\description{
Simulate matrices `ui` and `vi` of correlated random effects
from longitudinal hurdle model, where `ui` corresponds to the
zero part of the model, and `vi` corresponds to the non-zero part.
The number of rows is equal to the number of MCMC samples, the
number of columns is equal to the number of simulations used
in the inner Monte Carlo integration
}
