% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_mean.R
\name{marg_mean_q}
\alias{marg_mean_q}
\title{Summarize the posterior distribution of the 'marginal mean' at a give `newdata`}
\usage{
marg_mean_q(newdata, nsims, fit)
}
\arguments{
\item{newdata}{A data frame with covariates.}

\item{nsims}{The number of Monte Carlo samples used in the inner integration when calculating the marginal mean}

\item{fit}{The brmsfit object}
}
\value{
Quantiles 0.025, 0.5, 0.975 from a posterior sample of marginal mean, evaluated at `newdata`
}
\description{
Summarize the posterior distribution of the 'marginal mean' at a give `newdata`
}
